src/requirements/requirements_dev.txt
-r base.txt
django-debug-toolbar==1.8
sphinx
lxml
gunicorn
python-logstash
gevent
